refactor: extract inventory command service
